Remove compiler warnings

+++ common/cmd_usbd.c
@@ -545,7 +545,7 @@ static int process_data(struct usbd_ops *usbd)
 		/* Erase */
 		nand_cmd(0, offset, length, NULL);
 		/* Write */
-		sprintf(length, "%x", len);
+		sprintf(length, "%x", (unsigned int) len);
 		ret = nand_cmd(1, ramaddr, offset, length);
 		break;
 
@@ -567,7 +567,7 @@ static int process_data(struct usbd_ops *usbd)
 		/* Erase */
 		nand_cmd(0, offset, length, NULL);
 		/* Write */
-		sprintf(length, "%x", len);
+		sprintf(length, "%x", (unsigned int) len);
 		ret = nand_cmd(1, ramaddr, offset, length);
 		break;
